feat: rename multicodec pkg to codecregistry, leaving redirect
Fixes: https://github.com/ipld/go-ipld-prime/issues/259

I'm not sure this is possible since `DefaultRegistry` is not a pointer, so we
can't simply provide a redirect for it without breaking code that uses it. I've
left the dependent code intact here for now to demonstrate breakage.
